backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
PyMuPDF==1.24.10
python-docx==1.1.2
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
langchain==0.3.1
langchain-core==0.3.6
langchain-text-splitters==0.3.0
langchain-ollama==0.2.0
langchain-google-genai==2.0.0
google-generativeai==0.8.1
pydantic[email]==2.9.2
python-dotenv==1.0.1
rank-bm25==0.2.2
requests==2.32.3
beautifulsoup4==4.12.3
duckduckgo-search==6.3.7
